requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
dnspython==2.4.2 ; python_version >= '3.8' and python_version < '4.0'
filelock==3.12.2 ; python_version >= '3.7'
frozenlist==1.4.0 ; python_version >= '3.8'
graphql-core==3.2.3 ; python_version >= '3.6' and python_version < '4'
idna==3.4 ; python_version >= '3.5'
mongoengine==0.27.0
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.25.2 ; python_version >= '3.9'
pandas==2.0.3 ; python_version >= '3.8'
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.8.0 ; python_version >= '3.7'
pymongo==4.5.0
pypushdeer==0.0.3
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0 ; python_version >= '3.7'
scipy==1.11.2 ; python_version < '3.13' and python_version >= '3.9'
sgqlc==16.3 ; python_version >= '3.7' and python_version < '4.0'
shinny-structlog==0.0.4 ; python_version >= '3.6'
simplejson==3.19.1 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqsdk==3.4.6
tqsdk-ctpse==1.0.2 ; python_version >= '3'
tqsdk-sm==1.0.5 ; python_version >= '3'
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4 ; python_version >= '3.7'
websockets==11.0.3 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
